Break out and annotate AxImporter
Investigating porting the implementation, cleaning up the stub code.
