+++ src/System.Windows.Forms.Design/src/PublicAPI.Shipped.txt
@@ -243,18 +243,6 @@ System.ComponentModel.Design.Serialization.BasicDesignerLoader.SetBaseComponentC
 ~System.Drawing.Design.ToolboxItemCollection.this[int index].get -> System.Drawing.Design.ToolboxItem
 ~System.Drawing.Design.ToolboxItemCollection.ToolboxItemCollection(System.Drawing.Design.ToolboxItem[] value) -> void
 ~System.Drawing.Design.ToolboxItemCollection.ToolboxItemCollection(System.Drawing.Design.ToolboxItemCollection value) -> void
-~System.Windows.Forms.Design.AxImporter.AxImporter(System.Windows.Forms.Design.AxImporter.Options options) -> void
-~System.Windows.Forms.Design.AxImporter.IReferenceResolver.ResolveActiveXReference(System.Runtime.InteropServices.UCOMITypeLib typeLib) -> string
-~System.Windows.Forms.Design.AxImporter.IReferenceResolver.ResolveComReference(System.Reflection.AssemblyName name) -> string
-~System.Windows.Forms.Design.AxImporter.IReferenceResolver.ResolveComReference(System.Runtime.InteropServices.UCOMITypeLib typeLib) -> string
-~System.Windows.Forms.Design.AxImporter.IReferenceResolver.ResolveManagedReference(string assemName) -> string
-~System.Windows.Forms.Design.AxImporter.Options.keyContainer -> string
-~System.Windows.Forms.Design.AxImporter.Options.keyFile -> string
-~System.Windows.Forms.Design.AxImporter.Options.keyPair -> System.Reflection.StrongNameKeyPair
-~System.Windows.Forms.Design.AxImporter.Options.outputDirectory -> string
-~System.Windows.Forms.Design.AxImporter.Options.outputName -> string
-~System.Windows.Forms.Design.AxImporter.Options.publicKey -> byte[]
-~System.Windows.Forms.Design.AxImporter.Options.references -> System.Windows.Forms.Design.AxImporter.IReferenceResolver
 ~System.Windows.Forms.Design.Behavior.BehaviorServiceAdornerCollection.Add(System.Windows.Forms.Design.Behavior.Adorner value) -> int
 ~System.Windows.Forms.Design.Behavior.BehaviorServiceAdornerCollection.AddRange(System.Windows.Forms.Design.Behavior.Adorner[] value) -> void
 ~System.Windows.Forms.Design.Behavior.BehaviorServiceAdornerCollection.AddRange(System.Windows.Forms.Design.Behavior.BehaviorServiceAdornerCollection value) -> void
